+++ domain/src/main/java/com/cmc/domain/feature/spot/usecase/GetCategorySpotsUseCase.kt
@@ -9,7 +9,7 @@ class GetCategorySpotsUseCase @Inject constructor(
     private val spotRepository: SpotRepository
 ) {
     suspend operator fun invoke(
-        categoryId: Long,
+        categoryId: Int,
         latitude: Double,
         longitude: Double,
         sortBy: String,
